requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==2.6.1
arabic-reshaper==2.1.3
asgiref==3.5.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
auditlog3==1.0.1
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
boto==2.49.0
boto3==1.20.25
botocore==1.23.25
cached-property==1.5.2
celery==4.4.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.0.9
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==38.0.1
cssselect2==0.5.0
cycler==0.11.0
daphne==3.0.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.10
django-allauth==0.51.0
django-anymail==8.5
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-autoslug==1.9.8
django-braces==1.15.0
django-celery-beat==1.5.0
django-ckeditor==6.5.1
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-dbbackup==4.0.2
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-environ==0.10.0
django-extensions==3.1.5
django-htmx==1.9.0
django-imagekit==4.1.0
django-jet-4==0.0.1
django-js-asset==2.0.0
django-mailjet==0.3.1
django-multiselectfield==0.1.12
django-nine==0.2.5
django-rest-auth==0.9.5
django-sendgrid-v5==1.2.0
django-storages==1.12.3
django-templated-mail==1.1.1
django-timezone-field==5.0
django-user-agents==0.4.0
django-video-encoding==1.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
elasticsearch==7.13.4
elasticsearch-dsl==7.4.0
ffmpeg-python==0.2.0
fonttools==4.38.0
frozenlist==1.3.0
future==0.18.2
geoip2==4.5.0
graphviz==0.20
gunicorn==20.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.11.3
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
kiwisolver==1.4.4
kombu==4.6.11
ldap3==2.9.1
lxml==4.8.0
mailjet-rest==1.3.4
Markdown @ git+https://github.com/Python-Markdown/markdown.git@965c06c398a394b1da06651bb4ed62be7609a946
MarkupSafe==2.1.1
matplotlib==3.6.2
maxminddb==2.2.0
moviepy==1.0.3
msgpack==1.0.4
multidict==6.0.2
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.4.60
packaging==21.3
pilkit==2.0
Pillow==8.4.0
proglog==0.1.10
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydirectory==0.1.15
pydot==1.4.2
Pygments==2.2.0
PyJWT==2.4.0
pyOpenSSL==22.1.0
pyparsing==3.0.7
PyPDF3==1.0.6
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==1.0.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==4.3.4
reportlab==3.6.9
requests==2.26.0
requests-oauthlib==1.3.1
s3transfer==0.5.0
sendgrid==6.9.7
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==3.5.0
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.63.1
twilio==7.7.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.0.1
tzdata==2023.3
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.7
user-agents==2.2.0
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.14.0
xhtml2pdf==0.2.6
yarl==1.7.2
zipp==3.7.0
zope.interface==5.5.0
